updated 'Budget' page to include modal data (for sending to piechart & database)
